moved some files and rewrote some url building code
