+++ projects/ui/src/Components/App.tsx
@@ -5,8 +5,8 @@ import { AppUtilsContextProvider } from "../Context/AppUtilsContext";
 import { defaultTheme, globalStyles } from "../Styles";
 import { mantineThemeOverride } from "../Styles/global-styles/mantine-theme";
 import PortalServerTypeChecker from "../Utility/PortalServerTypeChecker";
-import PostLoginRedirectHandler from "../Utility/PostLoginRedirectHandler";
-import SessionExpiryHandler from "../Utility/SessionExpiryHandler";
+import PostLoginRedirectHandler from "../Utility/login/PostLoginRedirectHandler";
+import SessionExpiryHandler from "../Utility/login/SessionExpiryHandler";
 import UserChecker from "../Utility/UserChecker";
 import AppContent from "./AppContent";
